requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
httplib2==0.20.2
idna==3.3
psycopg2-binary==2.9.3
protobuf==3.19.3
pyparsing==3.0.6
requests==2.27.1
urllib3==1.26.8
git+https://github.com/gentilnegocios/tape-py.git#egg=pytape
